+++ CHANGELOG.md
@@ -6,9 +6,24 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to
 [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Added the following methods to `JupiterOneClient`:
+
+  ```ts
+  const client = await new JupiterOneClient(options).init();
+
+  await client.integrationInstances.list();
+  await client.integrationInstances.get(id);
+  await client.integrationInstances.create(instance);
+  await client.integrationInstances.update(id, update);
+  await client.integrationInstances.delete(id);
+  ```
+
 ## 0.23.6
 
-- Replace deleteEntity with deleteEntityV2 
-- Add typings and resolve typing errors 
-- Remove entity property in `uploadGraphObjectsForDeleteSyncJob` 
-
+- Replace deleteEntity with deleteEntityV2
+- Add typings and resolve typing errors
+- Remove entity property in `uploadGraphObjectsForDeleteSyncJob`